Support more types of requests by fixing encoding and which parameters are generated.
To support different varieties of GETs with query strings, POSTs with
and without query strings, and either case with parameters in query
strings or in an Authorization header, we need two changes.

First, control encoding based on the type of string generated. In an
Authorization header, parameters are left unencoded. When included in
the query string, they should be produced in encoded form. This needs
to be handled carefully since we use the same methods to generate the
signature base string and the output string, but the signature base
string always needs parameters to be URL encoded.

Second, only include oauth parameters in generated string if the
request is for the Authorization.