Fix reuse of nonces.
The way nonces were being generated could result in the same nonce being reused
if many nonces are generated per second. This could occur because the nonces are
created using a pseudorandom number and the current timestamp in seconds. The
timestamp guarantees no reuse of values over a long period of time, but allows
reuse within a period of 1 second. The pseudorandom number should account for
that issue, but for some reason the libtwitcurl code was using % 1000, resulting
in at most 1000 unique values per second. The birthday paradox makes a collision
fairly likely with only a small number of calls per second.

The fix to this just removes the %, using the rand() value directly. This should
make it nearly impossible to generate a duplicate given the way rand() is
usually implemented. To verify this, we also add a test that checks nonces over
a large number of quickly-generated signatures. This easily fails when the mod
is included.

To support this change, this also adds a small static method for resetting the
initialization state. This is necessary since some tests need to use fixed
values whereas some, such as this test, need to be sure they are using the
normal code path. This method is public, but prefixed with __ to indicate it
shouldn't be considered part of the public interface.

Fixes #13.